Fix errors spliting agent.py
