+++ framework/scripts/agent_groups.py
@@ -15,6 +15,7 @@ path.append(dirname(argv[0]) + '/../framework')  # It is necessary to import Waz
 try:
     from wazuh import Wazuh
     from wazuh.agent import Agent
+    import wazuh.group as group
     from wazuh.exception import WazuhException
 except Exception as e:
     print("Error importing 'Wazuh' package.\n\n{0}\n".format(e))
@@ -39,7 +40,7 @@ def signal_handler(n_signal, frame):
 
 
 def show_groups():
-    groups_data = Agent.get_all_groups(limit=None)
+    groups_data = group.get_all_groups(limit=None)
 
     print("Groups ({0}):".format(groups_data['totalItems']))
     for g in groups_data['items']:
@@ -54,7 +55,7 @@ def show_group(agent_id):
 
 
 def show_agents_with_group(group_id):
-    agents_data = Agent.get_agent_group(group_id, limit=0)
+    agents_data = group.get_agent_group(group_id, limit=0)
 
     if agents_data['totalItems'] == 0:
         print("No agents found in group '{0}'.".format(group_id))
@@ -65,7 +66,7 @@ def show_agents_with_group(group_id):
 
 
 def show_group_files(group_id):
-    data = Agent.get_group_files(group_id)
+    data = group.get_group_files(group_id)
     print("{0} files for '{1}' group:".format(data['totalItems'], group_id))
 
     longest_name = 0
@@ -86,7 +87,7 @@ def unset_group(agent_id, quiet=False):
         ans = 'y'
 
     if ans.lower() == 'y':
-        msg = Agent.unset_group(agent_id)
+        msg = group.unset_group(agent_id)
     else:
         msg = "Cancelled."
 
@@ -101,7 +102,7 @@ def remove_group(group_id, quiet=False):
         ans = 'y'
 
     if ans.lower() == 'y':
-        data = Agent.remove_group(group_id)
+        data = group.remove_group(group_id)
         msg = data['msg']
         if not data['affected_agents']:
             msg += "\nNo affected agents."
@@ -121,7 +122,7 @@ def set_group(agent_id, group_id, quiet=False):
         ans = 'y'
 
     if ans.lower() == 'y':
-        msg = Agent.set_group(agent_id, group_id)
+        msg = group.set_group(agent_id, group_id)
     else:
         msg = "Cancelled."
 
@@ -136,7 +137,7 @@ def create_group(group_id, quiet=False):
         ans = 'y'
 
     if ans.lower() == 'y':
-        msg = Agent.create_group(group_id)
+        msg = group.create_group(group_id)
     else:
         msg = "Cancelled."
