+++ framework/wazuh/exception.py
@@ -92,6 +92,7 @@ class WazuhException(Exception):
         1725: 'Error registering a new agent',
         1726: 'Ossec authd is not running',
         1727: 'Error listing group files',
+        1728: 'Error removing group',
 
         # Manager:
